Extracting query params form to its own component
